Add print help command and remove CommandCLI
